organize and cleanup
