+++ src/custom_lib/DOM_manipulator/Styles.css.ts
