+++ src/custom_lib/DOM_manipulator/utils/util.ts
